requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
black==23.1.0
Django==4.1.7
PyYAML==6.0
djangorestframework==3.14.0
django-filter==22.1